+++ CHANGELOG.md
@@ -1,3 +1,36 @@
+## [Release 2.41.1](https://github.com/aws-amplify/amplify-android/releases/tag/release_v2.41.1)
+
+### Bug Fixes
+- **auth:** Return signed-out session for User Pool-only apps ([#3403](https://github.com/aws-amplify/amplify-android/issues/3403))
+- **foundation:** re-throw CancellationException from resultCatching ([#3408](https://github.com/aws-amplify/amplify-android/issues/3408))
+- **api:** fail fast on AppSync subscription connection failure ([#3411](https://github.com/aws-amplify/amplify-android/issues/3411))
+
+[See all changes between 2.41.0 and 2.41.1](https://github.com/aws-amplify/amplify-android/compare/release_v2.41.0...release_v2.41.1)
+
+## [Release 2.41.0](https://github.com/aws-amplify/amplify-android/releases/tag/release_v2.41.0)
+
+### Features
+- **predictions:** report a distinct exception when a liveness session is interrupted ([#3396](https://github.com/aws-amplify/amplify-android/issues/3396))
+
+### Bug Fixes
+- **api:** percent-encode semicolons in REST query parameter values ([#3387](https://github.com/aws-amplify/amplify-android/issues/3387))
+- **all:** resolve UnsafeOptInUsage and UseKtx errors from newer androidx lint ([#3395](https://github.com/aws-amplify/amplify-android/issues/3395))
+- **auth:** Recover device metadata keyed by a pre-2.30.3 username ([#3393](https://github.com/aws-amplify/amplify-android/issues/3393))
+
+[See all changes between 2.40.0 and 2.41.0](https://github.com/aws-amplify/amplify-android/compare/release_v2.40.0...release_v2.41.0)
+
+## [Release 2.40.0](https://github.com/aws-amplify/amplify-android/releases/tag/release_v2.40.0)
+
+### Features
+- **analytics:** add Event Enrichment client module ([#3337](https://github.com/aws-amplify/amplify-android/issues/3337))
+- **notifications:** add Amplify Connect client for Customer Profiles ([#3341](https://github.com/aws-amplify/amplify-android/issues/3341))
+
+
+### Bug Fixes
+- **all:** Publish dependencies under proper scoping ([#3362](https://github.com/aws-amplify/amplify-android/issues/3362))
+
+[See all changes between 2.39.0 and 2.40.0](https://github.com/aws-amplify/amplify-android/compare/release_v2.39.0...release_v2.40.0)
+
 ## [Release 2.39.0](https://github.com/aws-amplify/amplify-android/releases/tag/release_v2.39.0)
 
 ### Features